Restructure validation schema into utils
